Client
Added validation if card can be trown on table of defend attacking card.
Game is not working due to server API not giving table responses after the first pair is done and there is ocasional connection with server bug/crash which does not cause client to stop working, just does not send/receive jsons.